joben-backend/sql_app/requirements.txt
aiosmtplib==2.0.2
altgraph==0.17.2
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.3
fastapi-mail==1.4.1
future==0.18.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.4
jwt==1.3.1
macholib==1.15.2
MarkupSafe==2.1.5
outcome==1.3.0.post0
passlib==1.7.4
pip==24.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydantic-settings==2.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.37.0
wsproto==1.2.0
